Hom.css
There was some error in Git, that's why I am generating a PR in this way. For reference I am solving "Enhancement: Add hover effect to the main heading for a more dynamic appearance #310" (https://github.com/Vin205/Enyanjyoti/issues/310). I was assigned to work on it and here is my contribution.